Initial implementation of proxy classes for the Rebtel SDK.
Initial implementation of proxy classes for the Rebtel SDK,
corresponding to the native classes RebtelSDK, REBClient and
REBCall. This enables basic incoming and outgoing calls.

Incoming calls are currently only supported if the application is in the
foreground, due to the required methods for the REBClientDelegate
protocol to present local notifications if the application is in the
background. More specifically,
-[REBClientDelegate client:localNotificationForIncomingCallFromUser:]
must be implemented for that to work.

Adding prepare_rebtel_sdk.py to download the Rebtel SDK package,
including the iOS Framework to be linked against.

Adding prepare_rebtel_sdk.py as part of the build.py build process.